\name{EQUAL_insert_signature_data}
\alias{EQUAL_insert_signature_data}
\title{
Insert digital signature for data
}
\description{
"Insert digital signature for data using the private key generated by
\code{\link[=EQUAL_encrypt_generate_keys]{EQUAL_encrypt_generate_keys()}}
function and \pkg{openssl}. This uses the SHA384 algorithm for the hash
function."
}
\usage{
EQUAL_insert_signature_data(data, private_key_folder, key_name)
}
\arguments{
\item{data}{Data for which signature must be inserted}
\item{private_key_folder}{Location of the private key}
\item{key_name}{Name of the private key}
}
\value{
"\item{path_to_signed_file}{path to the signed data which is stored in
a file}"
\item{signature}{signature}
}
\references{
\url{https://sites.google.com/view/equal-group/home}
}
\author{
Kurinchi Gurusamy
}
\note{
"This is part of a suite of functions required to allow encrypting and
decrypting whole files and encrypting and decrypting columns of data
programs to run. This is unlikely to be used as a stand alone function."
}
\seealso{
\code{\link[=EQUAL_encrypt_generate_keys]{EQUAL_encrypt_generate_keys()}}
\code{\link[openssl:signature_create]{openssl::signature_create()}}
}
\concept{EQUALencrypt}
\examples{
library(openssl)
# Encryption keys ####
test_folder <- tempfile(pattern = "folder_")
public_key_folder <- paste0(test_folder, "/public_key_folder")
private_key_folder <- paste0(test_folder, "/private_key_folder")
dir.create(test_folder)
dir.create(public_key_folder)
dir.create(private_key_folder)
encryption_keys <- EQUAL_encrypt_generate_keys(
  public_key_folder = public_key_folder,
  private_key_folder = private_key_folder,
  key_name = "encryption_key.txt")
# Data ####
data <- lapply(1:3, function(x) {
  mean = sample(1:100, 1, replace = FALSE)
  sd = sample(1:100, 1, replace = FALSE)
  rnorm(100, mean = mean, sd = sd)
})
data <- do.call(cbind.data.frame, data)
colnames(data) <- paste0("v", formatC(1:3, width = 6, flag = "0"))
# Encrypt data ####
encrypted_data <- EQUAL_encrypt_data(data = data,
                                     public_key_folder = public_key_folder,
                                     key_name = "encryption_key.txt")
# Insert signature ####
signature <- EQUAL_insert_signature_data(data = encrypted_data,
private_key_folder = private_key_folder,
key_name = "encryption_key.txt")
}
